# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: RAGO API
  description: |
    RAGO is a local-first, privacy-first Go AI foundation with four core pillars:
    - **LLM**: Language model management and generation
    - **RAG**: Retrieval-augmented generation
    - **MCP**: Model Context Protocol for tool integration
    - **Agents**: Autonomous workflow orchestration
  version: 3.0.0
  contact:
    name: RAGO Support
    url: https://github.com/liliang-cn/rago
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:7127/api/v1
    description: Local development server
  - url: https://api.rago.ai/v1
    description: Production server
...
paths:
...
  /chat:
    post:
      tags: [Unified]
      summary: Chat with RAG and tools
      description: Process chat messages with optional RAG context and tool usage
      security:
        - bearerAuth: []
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Chat response generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /task:
    post:
      tags: [Unified]
      summary: Execute task
      description: Execute a complex task using agents, tools, and RAG
      security:
        - bearerAuth: []
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '200':
          description: Task executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    ChatRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
        use_rag:
          type: boolean
          default: true
        use_tools:
          type: boolean
          default: true
    ChatResponse:
      type: object
      properties:
        response:
          type: string
        sources:
          type: array
          items:
            type: string
        tools_used:
          type: array
          items:
            type: string
    TaskRequest:
      type: object
      required: [task]
      properties:
        task:
          type: string
        context:
          type: object
    TaskResponse:
      type: object
      properties:
        result:
          type: object
        steps:
          type: array
          items:
            type: object
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string